Refactor scripts directory
