+++ scripts/azureml-assets/validation_rules.json
@@ -3,8 +3,8 @@
     "ignore-file": [
       "azureml/assets/__init__.py:F401",
       "azureml/assets/environment/__init__.py:F401",
-      "azureml/assets/util/__init__.py:F401",
-      "azureml/assets/release/model_publish_utils/__init__.py:F401"
+      "azureml/assets/model/__init__.py:F401",
+      "azureml/assets/util/__init__.py:F401"
     ],
     "justifications": [
       "__init__.py - namespace level files excluded from validation"
@@ -14,9 +14,8 @@
     "exclude": [
       "azureml/assets/__init__.py",
       "azureml/assets/environment/__init__.py",
-      "azureml/assets/util/__init__.py",
-      "azureml/assets/release/model_publish_utils/__init__.py"
-
+      "azureml/assets/model/__init__.py",
+      "azureml/assets/util/__init__.py"
     ],
     "justifications": [
       "__init__.py - namespace level files excluded from validation"
