--- scripts/test-requirements.txt
@@ -1,1 +1,0 @@
-pytest==7.1.1
