+++ scripts/test/test_assets.py
